refactor: mover gerarMapaFontes para classe GeradorMapaCss
